t5_sum updated
