+++ jaseci_kit/jaseci_kit/modules/t5_sum/tests/test_data.py
@@ -1,3 +1,13 @@
-PAYLOAD = "The US has passed the peak on new coronavirus cases, President Donald Trump said and predicted that some states would reopen this month. The US has over 637,000 confirmed Covid-19 cases and over 30,826 deaths, the highest for any country in the world. At the daily White House coronavirus briefing on Wednesday, Trump said new guidelines to reopen the country would be announced on Thursday after he speaks to governors. We'll be the comeback kids, all of us, he said. We want to get our country back. The Trump administration has previously fixed May 1 as a possible date to reopen the world's largest economy, but the president said some states may be able to return to normalcy earlier than that."
+PAYLOAD = """The US has passed the peak on new coronavirus cases, President Donald
+        Trump said and predicted that some states would reopen this month. The US
+        has over 637,000 confirmed Covid-19 cases and over 30,826 deaths, the highest
+        for any country in the world. At the daily White House coronavirus briefing on
+        Wednesday, Trump said new guidelines to reopen the country would be announced
+        on Thursday after he speaks to governors. We'll be the comeback kids, all of
+        us, he said. We want to get our country back. The Trump administration has
+        previously fixed May 1 as a possible date to reopen the world's largest
+        economy, but the president said some states may be able to return to normalcy
+        earlier than that."""
 
-test_t5_sum_request = {"text": PAYLOAD}
+
+test_t5_sum_request = {"text": PAYLOAD, "min_length": 30, "max_length": 100}
